knowledge/pyproject.toml
[project]
name = "isabl-knowledge"
version = "0.1.0"
description = "Knowledge tree generator for the Isabl genomics platform"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
isabl-knowledge = "isabl_knowledge.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/isabl_knowledge"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
